Patients Page : Prescriptions

+++ website/views.py
@@ -78,30 +78,32 @@ def prescription():
 
 @views.route('/prescription_content_remove', methods = ['POST'])
 def prescription_content_remove():
-    data = json.loads(request.data)
-    pc = PrescriptionContent.query.filter_by(PRecord_ID = data['PRecord_ID']).first()
-    if pc:
-        db.session.delete(pc)
-        db.session.commit()
-    return redirect(url_for('views.home'))
+    if check_session()["Logged_In"] != False and check_session()["Role"] != "Patient" and request.method == 'POST':
+        data = json.loads(request.data)
+        pc = PrescriptionContent.query.filter_by(PRecord_ID = data['PRecord_ID']).first()
+        if pc:
+            db.session.delete(pc)
+            db.session.commit()
+        return redirect(url_for('views.home'))
 @views.route('/prescription_add_content', methods=['POST'])
 def prescription_add_content():
-    data = json.loads(request.data)
-    Prescription_ID = data["Prescription_ID"]
-    MedicineID = data["Medicine_ID"]
-    Box = data["Box"]
+    if check_session()["Logged_In"] != False and check_session()["Role"] != "Patient" and request.method == 'POST':
+        data = json.loads(request.data)
+        Prescription_ID = data["Prescription_ID"]
+        MedicineID = data["Medicine_ID"]
+        Box = data["Box"]
 
-    Prescription_ = Prescription.query.filter_by(Prescription_ID= Prescription_ID).first()
-    Medicine_ = Medicine.query.filter_by(Medicine_ID = MedicineID).first()
-    if Prescription_ and Medicine_:
-        PrescriptionContent_ = PrescriptionContent(
-                Prescription_ID = Prescription_ID,
-                Medicine_ID = Medicine_.Medicine_ID,
-                Box = Box
-        )
-        db.session.add(PrescriptionContent_)
-        db.session.commit()
-    return redirect(url_for('views.home'))
+        Prescription_ = Prescription.query.filter_by(Prescription_ID= Prescription_ID).first()
+        Medicine_ = Medicine.query.filter_by(Medicine_ID = MedicineID).first()
+        if Prescription_ and Medicine_:
+            PrescriptionContent_ = PrescriptionContent(
+                    Prescription_ID = Prescription_ID,
+                    Medicine_ID = Medicine_.Medicine_ID,
+                    Box = Box
+            )
+            db.session.add(PrescriptionContent_)
+            db.session.commit()
+        return redirect(url_for('views.home'))
 
 
 #delete schedule
